+++ src/main/java/com/casaleo/sistema_pos/repositories/FacturaRepository.java
@@ -6,11 +6,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
+import java.math.BigDecimal;
 
 public interface FacturaRepository extends JpaRepository<Factura, Integer> {
 
     // Pendientes
-    List<Factura> findByCliente_IdAndSaldoGreaterThan(Long clienteId, Double saldo);
+    List<Factura> findByCliente_IdAndSaldoGreaterThan(Integer clienteId, BigDecimal saldo);
 
     // ✅ Búsqueda por N° factura o por datos del cliente
     Page<Factura> findByNumeroFacturaContainingIgnoreCaseOrCliente_NombreContainingIgnoreCaseOrCliente_RucContainingIgnoreCase(
